Log the credential name and app ID used when opening a Google Play API session.
Until now, if an app ID was used with credentials that do not provide access to
that app, we would show a relatively unhelpful message from the Google Play API,
i.e. "No application was found for the given package name."

Before this point, the application ID isn't actually shown — it's checked for
consistency, but the value isn't logged until we actually start to upload APKs.

Now by showing the credential name and application ID, it should make it clearer
to the user which account is in use, and therefore why an error occurred.