Improve uploading of ProGuard mapping files for multiple APKs.
Also removed the need for relative paths, as this is more complicated to
configure, and for all situations that I could find, is unnecessary.

Now, if we find a single mapping file, that will be associated with all
the APKs being uploaded. Otherwise, we attempt to upload the mapping
file that corresponds to each APK by relying on the fact that the
Android Gradle Plugin currently writes the mapping files into
directories which have the same naming pattern as the output directories
for the APKs themselves.