Remove lots of logic for assigning APKs to a release track.
The backend *finally* does this on our behalf, meaning we don't have to
do loads of lookups and altering of various other tracks.

It's also a lot more flexible in general, allows staged rollouts for all
types of tracks, and even allows reducing the rollout percentage.

With this, it looks like the awful virtual "rollout" track is no longer
something we need to care about.

Error messages also seem to be a bit more meaningful, so we can also get
rid of a bunch of manual checks on the client side.

Yay!