Create some basic fake HttpResponse handlers
Using the built-in MockHttpTransport, we're able to create a
LowLevelHttpRequest which can return a LowLevelHttpResponse. That allows us
to set up a kind of test director where we can specify what response we want
to see for a specific request URL.

The basic full end-to-end test requires 7 HTTP requests to be mocked in order
to simulate a successful upload.