Increase connection timeouts as AAB uploads can be very large.
The device-specific bundle delivered to phones can be up to 150MB, so
the actual bundle uploaded to Google Play can be many times larger.

Joe was getting timeout exceptions when uploading ~100MB bundle files,
with the default timeout (which seems to be 20 seconds), and was seeing
upload times of 2.5 minutes after fixing the timeout, so increasing the
timeout will definitely be required (and it's presumably been a problem
for users in the past, not that anybody has actually complained…).

It's unclear what the read timeout means exactly; is it the amount of
time before receiving an HTTP response (i.e. rather than being a TCP
connection-level timeout), meaning that we don't "read" anything from
the HTTP server while we're uploading a file?

Anyway, now we have a one-hour read timeout, which I think should cover
most people's bundle sizes and internet connection speeds.  I also think
having a defined timeout is better than having an infinite timeout, as
most Jenkins users probably don't have a job timeout configured.