Rename fields, and add defaults, for ReleaseTrackAssignmentBuilder.
Changes were made in a backwards compatible way:
- `apkFilesPattern` was renamed to `filesPattern`
- `rolloutPercentage` was renamed to `rolloutPercent`

The latter is now numeric, and so can be used more easily from
Pipelines, without having to quote the percentage as a string.

I tested Freestyle jobs with the old configuration, Pipelines with the
old field names, and checked the Snippet Generator.

The `filesPattern`, `rolloutPercent`, and `trackName` fields now all
have default values, so the build step can potentially be used like:

  androidApkMove googlePlayCredentialsId: 'gp', fromVersionCode: false

Which would move any AABs or APKs found to the production track.