Allow applying expansion files from the current upload to multiple APKs.
i.e. if you have a new main or patch expansion file, and want to apply
that same file to multiple APKs being uploaded, previously you would
have to create multiple copies of the same expansion file and upload
them all.

Now you can name the expansion file to be uploaded according to the
lowest versionCode that you're uploading. That expansion file will then
be uploaded, and applied to the APKs with higher versionCodes that were
uploaded in the same build.