Revert to using a string field for rollout percentage.
In the latest release, I deprecated the `rolloutPercentage` string field
and replaced it with the numeric `rolloutPercent` field in order to make
Pipeline configuration a bit less awkward, e.g.

  `rolloutPercent: 12.34` instead of `rolloutPercentage: '12.34'`

However, while I tested build parameters with other fields, I didn't do
so for rollout percentage.  This means that if you enter a variable as
the rollout percentage, e.g. `${ROLLOUT_PCT}`, Jenkins would attempt to
coerce this to a Double value when saving the job, which apparently does
not crash, but instead returns `null`.

Since we interpret `null` to mean "the default value", entering a
variable into the rollout percentage field, saving the job, and then
opening the job config again would show "100.0" as the value.  Oops.

Now we've returned to `rolloutPercentage` being the default field, and
have marked `rolloutPercent` as deprecated.

However, we remain backwards compatible, and the `rolloutPercent` field
persisted for Freestyle jobs will be migrated back to the string value.
You can also continue to use the slightly more convenient numeric syntax
for Pipeline jobs — either field is accepted, with `rolloutPercentage`
taking precedence.