Acknowledge the vague "bundle installation warning" for bundle uploads.
A user was having their upload rejected by Google Play when using this
plugin with a bundle file of around 200MB:

  The installation of the app bundle may be too large and trigger user
  warning on some devices, and this needs to be explicitly acknowledged
  in the request.

Indeed, the API provides an `ackBundleInstallationWarning` flag in order
to acknowledge this when uploading a bundle.  However, I wasn't able to
find any other reference to this "user warning" in the Google Play docs.

There _is_ a reference to a warning dialog, though it explicitly relates
to APKs rather than bundles, and encourages developers to switch to AAB:
https://support.google.com/googleplay/android-developer/answer/2481797

  If the compressed size of your APK at the time of download (including
  expansion files) is larger than 100MB, users see a warning dialog
  suggesting that they use Wi-Fi to download your app.

Now we acknowledge this flag for all bundles, regardless of file size,
as setting the flag even for very small bundles seems to work fine.

i.e. users of this plugin who wish to upload large bundles should not
have to change their job config, and uploading should now succeed.

See https://issues.jenkins-ci.org/browse/JENKINS-62389 for details.