Retain release notes when assigning files to a new release track.
This isn't done automatically when updating a release track. So we now
search for an existing track that contains one of the files we want to
move, and apply the associated release notes, if any, to the new track.

This isn't 100% foolproof, as it's possible to have the same version
codes assigned to multiple tracks, so it's possible for there to be
differing release notes in multiple tracks, for the same versionCodes.
i.e. you might not always get the release notes you wanted in the new
track, but I imagine that to be something of an edge case.