Remove 'production' as the default track name for the upload build step.
This is a breaking change, which will require users to explicitly
provide the track name.

This makes the plugin safer, as publishing to production by default,
e.g. in the absence of configuration, or due to a mistake when defining
a parameter, is probably unexpected behaviour for most users.

It might have been possible to differentiate between a `null` value and
an empty string configured for the `trackName` field, but as we've
changed how this field is interpreted and stored across releases of this
plugin, this wouldn't be reliable.

Better to make this breaking change, and make it clear to all users, old
and new, that they need to specify the track name.