Make the default value for percentage appear in config forms.
This makes it consistent with how the files pattern works, i.e. its
default value appears in the job config form, and the Pipeline snippet
generator.

Also made percentage values appear with the %-sign suffix, as it looks
a little bit nicer, and helps further reinfoce that it's a percentage
value and not a fraction, i.e. should be set to '50%' rather than '0.5'.